Follow the instructions to make a mask!
Supplies:
•	Duck-tape
•	Paper
•	Printer(8inch*11inch)
•	Scissors
•	Rubber Band or String (or something to hold to face)
•	Tissue
Instructions:
1.	Print out this PDF.

2.	Wrap another piece of paper (not this print out) in duck-tape on both sides once or twice.

3.	Place the page on top of the wrapped paper and cut the mold out of the wrapped paper. Do not forget to cut out the air hole and the side holes (this will require the duck-tape paper to be folded).

4.	Numbers 1, 2, 3, and 4 are marked on the mask mold as folding reference but do all the folding on the duck-tape mask. Fold 2 (the largest part) in half. Fold 1(the bottom part) under two and duck-tape the sides. Fold 3 down and fold the 4s down. Then duck-tape the 4s to the 3.

5.	Tie a rubber band or sting through side holes, and then place duck-tape around the edges of tissue and tape over the airhole. The tissue will need to be replace each time the mask is sterilized.

How to use: 
This mask can be sterilized with boiling water and reused multiple times (the duck-tape may change a little bit in boiling water). Due to duck-tape’s adhesion, I do not recommend placing the mask in a pan that is use for cooking, just pour the water over the mask in the sink or driveway.

Please Share this file with anyone who needs it and make as many masks as you please!